moving type to type file
