+++ apps/passport-server/src/services/generic-issuance/types.ts
@@ -1,4 +1,6 @@
+import { PipelineDefinition } from "@pcd/passport-interface";
 import { PipelineCapability } from "./capabilities/types";
+import { Pipeline, PipelineUser } from "./pipelines/types";
 
 /**
  * TODO:
@@ -19,4 +21,21 @@ export interface GenericIssuanceUser {
 export interface BasePipelineCapability {
   type: PipelineCapability;
   urlPath?: string; // set dynamically during application initialization
+}
+
+/**
+ * It's not always possible to start a {@link Pipeline} given a {@link PipelineDefinition}
+ * because a pipeline could be misconfigured.
+ *
+ * An {@link PipelineSlot} is used to represent a pair of {@link PipelineDefinition} and
+ * its corresponding {@link Pipeline} if one was able to be started.
+ */
+export interface PipelineSlot {
+  definition: PipelineDefinition;
+
+  // runtime information - intentionally ephemeral
+  instance?: Pipeline;
+  owner?: PipelineUser;
+  loadIncidentId?: string;
+  lastLoadDiscordMsgTimestamp?: Date;
 }
